backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pymongo>=4.5.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
werkzeug>=3.0.0
bcrypt>=4.0.0
